Recording of an updated rio, also Keychron K8
Because of a different keyboard (which I don't think matters, but it's worth
keeping in mind) I've done this as a fresh recording. I should go back and
recheck the Logi recording with rio on macOS to ensure that it's pretty much
in line.

The headline difference here seems to be that function keys are now working.